refactor(utils): Move content parser to utils package
- Relocate `getWholeParsedContent` and `getParsedReasoningContent` functions from ai-workspace-common to utils package
- Update import paths in affected files to use new utils location
- Export content parser functions in utils index file
